src/requirements.txt
discord.py==1.7.2
mysql-connector-python==8.0.18
requests